LPD-23201 Move ObjectFieldInfoFieldConverter to Object Info API module
