+++ modules/apps/object/object-info-api/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Object Info API
+Bundle-SymbolicName: com.liferay.object.info.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.object.info.field.converter
