+++ modules/apps/object/object-info-api/src/main/resources/com/liferay/object/info/field/converter/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
