+++ modules/apps/object/object-web/build.gradle
@@ -39,6 +39,7 @@ dependencies {
 	compileOnly project(":apps:notification:notification-api")
 	compileOnly project(":apps:object:object-admin-rest-api")
 	compileOnly project(":apps:object:object-api")
+	compileOnly project(":apps:object:object-info-api")
 	compileOnly project(":apps:object:object-rest-api")
 	compileOnly project(":apps:portal-configuration:portal-configuration-module-configuration-api")
 	compileOnly project(":apps:portal-security:portal-security-script-management-api")
